taotao/requirement.txt
tornado==4.3
mongoengine==0.10.0
pymongo==3.0.3
redis==2.10.5
bcrypt==2.0.0
elasticsearch==2.3.0
jieba==0.38
python-dateutil==2.5.2
pyjwt==1.4.0
pillow==2.8.0
requests==2.9.1
celery==3.1.23
qiniu==7.0.6
